Extract common code for change histories
