+++ ui/src/components/common/ChangeHistory/index.ts
@@ -7,4 +7,5 @@ export * from './DateRangeFilter';
 export * from './EmptyCell';
 export * from './FailedToLoadChangeHistory';
 export * from './LoadingChangeHistory';
+export * from './SectionTitle';
 export * from './StopsList';
