+++ ui/src/components/stop-registry/components/ChangeHistory/index.ts
@@ -1,0 +1,3 @@
+export * from './DataDiffFailedToLoadSection';
+export * from './DataDiffSectionLoading';
+export * from './types/PreviousStopPlaceChangeHistoryItem';
