+++ ui/src/components/stop-registry/components/ChangeHistory/types/PreviousStopPlaceChangeHistoryItem.ts
