+++ ui/src/components/stop-registry/stop-areas/change-history/components/DataDiffSections.tsx
@@ -2,16 +2,20 @@ import { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { StopPlaceChangeHistoryItem } from '../../../../../generated/graphql';
 import { GetUserNameById } from '../../../../../hooks';
-import { ChangeValueSections } from '../../../../common/ChangeHistory';
+import {
+  ChangeValueSections,
+  SectionTitle,
+} from '../../../../common/ChangeHistory';
+import {
+  DataDiffFailedToLoadSection,
+  DataDiffSectionLoading,
+} from '../../../components/ChangeHistory';
 import { useGetHistoricalStopAreaDetails } from '../queries';
 import {
   diffStopAreaBasicDetails,
   diffStopAreaStops,
   diffStopAreaTerminal,
 } from '../utils';
-import { DataDiffFailedToLoadSection } from './DataDiffFailedToLoadSection';
-import { DataDiffSectionLoading } from './DataDiffSectionLoading';
-import { SectionTitle } from './SectionTitle';
 
 type DataDiffSectionsProps = {
   readonly getUserNameById: GetUserNameById;
@@ -44,6 +48,14 @@ export const DataDiffSections: FC<DataDiffSectionsProps> = ({
       <DataDiffSectionLoading
         getUserNameById={getUserNameById}
         historyItem={historyItem}
+        sectionTitle={
+          <SectionTitle
+            historyItem={historyItem}
+            section={t(($) => $.stopAreaChangeHistory.sectionTitle)}
+          />
+        }
+        headerTestId="LoadingStopAreaData"
+        contentTestId="ChangeHistory::LoadingStopAreaData::Content"
       />
     );
   }
@@ -61,6 +73,14 @@ export const DataDiffSections: FC<DataDiffSectionsProps> = ({
             previousRefetch();
           }
         }}
+        sectionTitle={
+          <SectionTitle
+            historyItem={historyItem}
+            section={t(($) => $.stopAreaChangeHistory.sectionTitle)}
+          />
+        }
+        headerTestId="FailedToLoadStopAreaData"
+        retryButtonTestId="ChangeHistory::FailedToLoadStopAreaData::RetryButton"
       />
     );
   }
