+++ ui/src/components/stop-registry/stop-areas/change-history/components/StopAreaChangeHistoryFirstVersionHeaderRow.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from 'react-i18next';
 import { StopPlaceChangeHistoryItem } from '../../../../../generated/graphql';
 import { GetUserNameById } from '../../../../../hooks';
 import { ChangeHistoryItemSectionHeaderRow } from '../../../../common/ChangeHistory';
-import { SectionTitle } from './SectionTitle';
+import { SectionTitle } from '../../../../common/ChangeHistory/SectionTitle';
 
 const testIds = {
   // These will expand to: ChangeHistory::SectionHeader::${testId}
