--- ui/src/components/stop-registry/stop-areas/change-history/types/index.ts
@@ -1,1 +1,0 @@
-export * from './PreviousStopPlaceChangeHistoryItem';
