+++ ui/src/components/stop-registry/stop-areas/stop-area-details/components/StopAreaLatestChanges/LatestStopAreaChangeHistoryItem.tsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router';
 import { StopPlaceChangeHistoryItem } from '../../../../../../generated/graphql';
 import { Path, routeDetails } from '../../../../../../router/routeDetails';
 import { NoEarlierVersionExists } from '../../../../../common/ChangeHistory';
-import { PreviousStopPlaceChangeHistoryItem } from '../../../change-history/types';
+import { PreviousStopPlaceChangeHistoryItem } from '../../../../components/ChangeHistory';
 import { LatestStopAreaChangeDataDiff } from './LatestStopAreaChangeDataDiff';
 
 const testIds = {
