+++ ui/src/locales/en-US/common.json
@@ -1497,6 +1497,8 @@
     "failedToLoad": "Failed to load change history",
     "tryAgainButton": "Try again",
     "noChangedValues": "New version was saved, but nothing was changed.",
+    "versionLoading": "Resolving changed fields",
+    "sectionFailedToLoad": "Failed to load change details",
     "tableHeaders": {
       "name": "Name",
       "oldValue": "Old value",
@@ -1505,19 +1507,24 @@
       "validityEnd": "Valid to",
       "changedBy": "Changed by",
       "changed": "Changed"
+    },
+    "infoSpots": {
+      "posters": "Posters",
+      "added": "New info spot",
+      "updated": "Updated info spot",
+      "removed": "Removed info spot"
     }
   },
   "stopChangeHistory": {
     "title": "Change history | Stop {{publicCode}}",
     "titleWithName": "Change history | Stop  {{publicCode}} {{name}}",
     "titleLatest": "Change history",
+    "sectionTitle": "Stop",
     "showAll": "Show change history",
     "goBack": "Back",
     "firstVersion": "New stop created",
     "importedVersion": "Stop imported from Jore 3",
     "copiedVersion": "Stop copied from another version",
-    "failedToLoad": "Failed to load change details",
-    "versionLoading": "Resolving changed fields",
     "ownerDetails": "Owner details",
     "noChangesTitle": "Stop",
     "shelters": {
@@ -1527,11 +1534,7 @@
       "removed": "Removed shelter"
     },
     "infoSpots": {
-      "title": "Info Spots",
-      "posters": "Posters",
-      "added": "New Info Spot",
-      "updated": "Updated Info Spot",
-      "removed": "Removed Info Spot"
+      "title": "Info Spots"
     },
     "stopPlace": {
       "title": "Stop Area and Terminal",
