+++ ui/src/locales/fi-FI/common.json
@@ -1497,6 +1497,8 @@
     "failedToLoad": "Muutoshistorian lataus epäonnistui",
     "tryAgainButton": "Yritä uudelleen",
     "noChangedValues": "Uusi versio tallennettu, mutta siinä ei ole mitään muutoksia.",
+    "versionLoading": "Selvitetään muuttuneita kenttiä",
+    "sectionFailedToLoad": "Muuttuneiden kenttien selvitys epäonnistui",
     "tableHeaders": {
       "name": "Nimi",
       "oldValue": "Vanha arvo",
@@ -1505,19 +1507,24 @@
       "validityEnd": "Päättyy",
       "changedBy": "Päivittäjä",
       "changed": "Päivitetty"
+    },
+    "infoSpots": {
+      "posters": "Infotuotteet",
+      "added": "Uusi infopaikka",
+      "updated": "Päivitetty infopaikka",
+      "removed": "Poistettu infopaikka"
     }
   },
   "stopChangeHistory": {
     "title": "Muutoshistoria | Pysäkki {{publicCode}}",
     "titleWithName": "Muutoshistoria | Pysäkki {{publicCode}} {{name}}",
     "titleLatest": "Muutoshistoria",
+    "sectionTitle": "Pysäkki",
     "showAll": "Näytä muutoshistoria",
     "goBack": "Palaa",
     "firstVersion": "Uusi pysäkki luotu",
     "importedVersion": "Pysäkki tuotu Jore 3:sta",
     "copiedVersion": "Pysäkki kopioitu toisesta versiosta",
-    "failedToLoad": "Muuttuneiden kenttien selvitys epäonnistui",
-    "versionLoading": "Selvitetään muuttuneita kenttiä",
     "ownerDetails": "Omistajan tiedot",
     "noChangesTitle": "Pysäkki",
     "shelters": {
@@ -1527,11 +1534,7 @@
       "removed": "Poistettu katos"
     },
     "infoSpots": {
-      "title": "Infopaikat",
-      "posters": "Infotuotteet",
-      "added": "Uusi infopaikka",
-      "updated": "Päivitetty infopaikka",
-      "removed": "Poistettu infopaikka"
+      "title": "Infopaikat"
     },
     "stopPlace": {
       "title": "Pysäkkialue ja terminaali",
